+++ src/editor.rs
@@ -1,16 +1,14 @@
-use nih_plug::editor::Editor;
-use nih_plug_vizia::vizia::prelude::*;
-use nih_plug_vizia::{assets, create_vizia_editor, ViziaState, ViziaTheming};
-use rand::rngs::SmallRng;
-use rand::Rng;
-use realfft::num_complex::Complex;
-use rtrb::*;
-use std::collections::HashSet;
-use std::sync::mpsc::{self, Receiver, TryRecvError};
+use crate::{consts::*, AutomataParams};
+
+use std::sync::mpsc::{self, Sender};
 use std::sync::Arc;
 use std::thread;
 
-use crate::{AutomataParams, FILTER_WINDOW_SIZE};
+use nih_plug::editor::Editor;
+use nih_plug_vizia::vizia::prelude::*;
+use nih_plug_vizia::{assets, create_vizia_editor, ViziaState, ViziaTheming};
+use realfft::num_complex::Complex;
+use rtrb::*;
 
 #[derive(Lens)]
 struct Data {
@@ -94,113 +92,4 @@ pub(crate) fn create(
     });
 
     (cons, e)
-}
-
-fn game_loop(gui_reciever: Receiver<GUIEvent>, mut ir_producer: Producer<Complex<f32>>) {
-    let mut alive_cells =
-        HashSet::<(i32, i32)>::with_capacity(DEFAULT_IR_SPECTRUM_SIZE * DEFAULT_IR_SPECTRUM_SIZE);
-    let mut game_running = false;
-    let mut rng = rand::thread_rng();
-    build_random(&mut alive_cells, &mut rng, DEFAULT_IR_SPECTRUM_SIZE);
-
-    loop {
-        let message = gui_reciever.try_recv();
-        match message {
-            Ok(GUIEvent::PlayPause) => {
-                game_running = !game_running;
-            }
-            Ok(GUIEvent::Reset) => {
-                game_running = false;
-                alive_cells.drain();
-                build_random(&mut alive_cells, &mut rng, DEFAULT_IR_SPECTRUM_SIZE);
-            }
-            Err(TryRecvError::Empty) => {}
-            Err(TryRecvError::Disconnected) => panic!("gui disconnected from game loop"),
-        }
-
-        if game_running {
-            let mut dying: Vec<(i32, i32)> = Vec::new();
-            let mut born: Vec<(i32, i32)> = Vec::new();
-
-            for cell in alive_cells.iter() {
-                let neighbors = find_neighbors(&cell);
-                let mut living_neighbors: u8 = 0;
-                for neighbor in neighbors.iter() {
-                    if alive_cells.contains(neighbor) {
-                        living_neighbors += 1;
-                    } else if !born.contains(neighbor) {
-                        let neighbors_neighbors = find_neighbors(neighbor);
-                        let mut neighbors_living_neighbors: u8 = 0;
-                        for neighbor_neighbor in neighbors_neighbors.iter() {
-                            if alive_cells.contains(neighbor_neighbor) {
-                                neighbors_living_neighbors += 1;
-                            }
-                        }
-                        if neighbors_living_neighbors == 3 && !born.contains(neighbor) {
-                            born.push(*neighbor);
-                        }
-                    }
-                }
-                if living_neighbors > 3 || living_neighbors < 2 && !dying.contains(cell) {
-                    dying.push(*cell);
-                }
-            }
-            for cell in dying.iter() {
-                alive_cells.remove(cell);
-            }
-            for cell in born.iter() {
-                alive_cells.insert(*cell);
-            }
-            dying.clear();
-            born.clear();
-
-            let ir = build_ir(&alive_cells);
-            match ir_producer.write_chunk_uninit(DEFAULT_IR_SPECTRUM_SIZE) {
-                Ok(chunk) => {
-                    chunk.fill_from_iter(ir.into_iter());
-                }
-                Err(_) => {
-                    todo!();
-                }
-            }
-        }
-    }
-}
-
-fn find_neighbors(pos: &(i32, i32)) -> Vec<(i32, i32)> {
-    let mut neighbors: Vec<(i32, i32)> = Vec::new();
-    for x in -1..2 {
-        for y in -1..2 {
-            if x != 0 || y != 0 {
-                neighbors.push((pos.0 + x, pos.1 + y));
-            }
-        }
-    }
-    neighbors
-}
-
-pub fn build_random(board: &mut HashSet<(i32, i32)>, rng: &mut SmallRng, size: usize) {
-    for i in 0..size {
-        for j in 0..size {
-            if rng.gen_bool(0.5) {
-                board.insert((i as i32, j as i32));
-            }
-        }
-    }
-}
-
-pub fn build_ir(board: &HashSet<(i32, i32)>, size: usize) -> Vec<Complex<f32>> {
-    let mut out = vec![Complex::<f32> { re: 0.0, im: 0.0 }; size];
-    let mut i = 0;
-    for cell in board.iter() {
-        if i % 2 == 0 {
-            out[cell.0 as usize].re += 1 as f32 / size as f32;
-            out[cell.1 as usize].im += 1 as f32 / size as f32;
-        } else {
-            out[cell.0 as usize].re -= 1 as f32 / size as f32;
-            out[cell.1 as usize].im -= 1 as f32 / size as f32;
-        }
-        i += 1;
-    }
-    out
 }
